💡 refactor|CoverEffect|Use AE from item for CE config
Previously, the known config state was controlled by the Cover Effect object. This made it challenging to keep in sync with the underlying active effect data on the effects item.

Changed so CE config now pulled from the stored AE on the item. Simpler to update b/c can use the AE config to update directly.