🐛 fix|CoverEffect|Make _coverAttackers static
Cover attackers is really not dependent on the current token, other than to eliminate it as a possible attacker so it doesn't attack itself. Also, cover effects previously did not always respect the settings for application based on attacker.

Made `coverAttackers` static. Now setting cover effects will first check if the attackers are the same as for cover types, and modify the cover types accordingly for purposes of setting cover effects. This respects targeting and combat settings where the attackers might differ.