💡 refactor|TokenCover|Move token cover methods and getters to class
The number of added methods and getters to TokenClass was getting extreme. No clear way to extend class instances using submethods, like `_token.tokencover.myMethod` because it is not easy to pass along the token as `this` to the method.

Instead, create a `TokenCover` class to handle it. Use a getter on `Token.prototype` to make it accessible using that same path: `_token.tokencover.myMethod`. Now everything is behind the class, reducing naming clashes.