🐛 fix|CoverEffects|Fixes to ensure the dialog is awaited
The cover effects list was not getting re-rendered after a dialog, because the dialog was not properly awaiting the outcome due to a callback issue.

Somewhat simplified the dialog callback so now it awaits the outcome and returns the entire html. To make it more generic, it returns the button pressed and the html.