🐛 fix|Settings|Trigger cover type and effect change on settings change
If user changes when to apply cover type or effect, we want that to result in an immediate change to the tokens. So trigger a forced update of token cover. Forced so that attackers will be recalculated and refresh will occur even if the setting change only affected the refresh.