🐛 fix|Macros|Update cover macros
Remove the cover set macro, as it conflicts with how token cover now works. The macro could be used to force cover for a user, but it would only work for the user and only until something changes.

Update the debug cover macro with the new location of the token coverCalculator. Update other methods to reflect tokencover and the new coverCalculator location.