🎸 feat|CoverCalc|Add method to clone an attacker token
For modules like walled templates or tactical grid, it is useful to have a cover calculator based on a token but at a different token location. Add method to clone a token and return a cover calculator, so cover can be calculated based on this new location for the attacker.