💡 refactor|CoverObject|Track cover types and effects in TokenCover
Cover types cannot be tracked by icon file name alone, as multiple effects/statuses could use the same icon. Instead, track in TokenCover.

Cover effects are unique to the active effect and should be tracked that way. Use a getter to track in TokenCover.

This simplifies CoverType and CoverEffect as they need less logic to function.