🐛 fix|UniqueEffect|Drop use of sockets for UniqueEffect
As with TM, rely on the GM to create new cover documents.
For regions, all users will be notified, so use the first GM.
Otherwise, either the GM is the one making the change or it is a local change for which sockets is not required.

Still need sockets for GM confirm dialog.