🎸 feat|CoverRegion|Add option to apply to attackers within region
Checkbox to apply to attackers within region.
- Disabled: If defender is within region, it gains cover from the region. If not within, no cover.
- Enabled: If attacker is within region, defender gains cover regardless of defender location.

Consider forest. Cover might apply if attacker is in forest and defender is outside in the meadow, depending of course on how thick the forest. Versus token from above, where attacker in forest can shoot clearly at a flying token but the flying token might have trouble hitting things in the forest.